[MOB-16085] -- Move nested class from GeminioRecipe class
