+++ plugins/hh-geminio/src/main/kotlin/ru/hh/plugins/geminio/model/RequiredParams.kt
@@ -1,0 +1,7 @@
+package ru.hh.plugins.geminio.model
+
+
+data class RequiredParams(
+    val name: String,
+    val description: String,
+)
