+++ plugins/hh-geminio/src/test/kotlin/ru/hh/plugins/geminio/model/mapping/GeminioRecipeBooleanExpressionSpec.kt
@@ -4,9 +4,9 @@ import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldStartWith
-import ru.hh.plugins.geminio.model.GeminioRecipe.RecipeExpression.Command
-import ru.hh.plugins.geminio.model.GeminioRecipe.RecipeExpression.Command.ReturnFalse
-import ru.hh.plugins.geminio.model.GeminioRecipe.RecipeExpression.Command.ReturnTrue
+import ru.hh.plugins.geminio.model.RecipeExpression.Command
+import ru.hh.plugins.geminio.model.RecipeExpression.Command.ReturnFalse
+import ru.hh.plugins.geminio.model.RecipeExpression.Command.ReturnTrue
 import ru.hh.plugins.geminio.tests_helpers.GeminioExpressionUtils.createParametersMap
 import ru.hh.plugins.geminio.tests_helpers.GeminioExpressionUtils.toExpression
